fix warning

+++ src/tinflate.c
@@ -467,7 +467,7 @@ static int tinf_inflate_block_data(TINF_DATA *d, TINF_TREE *lt, TINF_TREE *dt)
 #endif // !NO_DICT
         {
             /* catch trying to point before the start of dest buffer */
-            if (offs > d->dest - d->destStart) {
+            if ((long)offs > d->dest - d->destStart) {
                 return TINF_DATA_ERROR;
             }
             d->lzOff = -offs;
